Re-add deleted directories to package zips
- Restores support for empty directories in app packages

[#147159371]

+++ lib/cloud_controller/packager/local_bits_packer.rb
@@ -8,27 +8,17 @@ module CloudController
       def send_package_to_blobstore(blobstore_key, uploaded_package_zip, cached_files_fingerprints)
         matched_resources = CloudController::Blobstore::FingerprintsCollection.new(cached_files_fingerprints)
 
-        FileUtils.chmod('u+w', uploaded_package_zip)
-
         Dir.mktmpdir('local_bits_packer', tmp_dir) do |root_path|
           app_package_zip = File.join(root_path, 'copied_app_package.zip')
-          FileUtils.cp(uploaded_package_zip, app_package_zip)
-
           app_packager = AppPackager.new(app_package_zip)
 
-          # unzip app contents & upload to blobstore
-          app_contents_path = File.join(root_path, 'application_contents')
-          FileUtils.mkdir(app_contents_path)
-          app_packager.unzip(app_contents_path)
-          global_app_bits_cache.cp_r_to_blobstore(app_contents_path)
-
-          # download & append cached resources from blobstore
-          cached_resources_dir = File.join(root_path, 'cached_resources_dir')
-          FileUtils.mkdir(cached_resources_dir)
-          matched_resources.each do |local_destination, file_sha, mode|
-            global_app_bits_cache.download_from_blobstore(file_sha, File.join(cached_resources_dir, local_destination), mode: mode)
+          if package_zip_exists?(uploaded_package_zip)
+            FileUtils.chmod('u+w', uploaded_package_zip)
+            FileUtils.cp(uploaded_package_zip, app_package_zip)
+            populate_resource_cache(app_packager, root_path)
           end
-          app_packager.append_dir_contents(cached_resources_dir)
+
+          append_matched_resources(app_packager, matched_resources, root_path)
 
           app_packager.fix_subdir_permissions
           validate_size!(app_packager)
@@ -43,6 +33,26 @@ module CloudController
       end
 
       private
+
+      def package_zip_exists?(package_zip)
+        package_zip && File.exist?(package_zip)
+      end
+
+      def populate_resource_cache(app_packager, root_path)
+        app_contents_path = File.join(root_path, 'application_contents')
+        FileUtils.mkdir(app_contents_path)
+        app_packager.unzip(app_contents_path)
+        global_app_bits_cache.cp_r_to_blobstore(app_contents_path)
+      end
+
+      def append_matched_resources(app_packager, matched_resources, root_path)
+        cached_resources_dir = File.join(root_path, 'cached_resources_dir')
+        FileUtils.mkdir(cached_resources_dir)
+        matched_resources.each do |local_destination, file_sha, mode|
+          global_app_bits_cache.download_from_blobstore(file_sha, File.join(cached_resources_dir, local_destination), mode: mode)
+        end
+        app_packager.append_dir_contents(cached_resources_dir)
+      end
 
       def validate_size!(app_packager)
         return unless max_package_size
